pyproject.toml
[tool.poetry]
name = "crispr"
version = "0.1.0"
description = ""
authors = ["evgenii.basmov <coykto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
biopython = "^1.81"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
